Fix return type annotation for task_wait_n function to List[float]

+++ python_async_function/4-tasks.py
@@ -4,7 +4,7 @@ from typing import List
 task_wait_random = __import__('3-tasks').task_wait_random
 
 
-async def task_wait_n(n: int, max_delay: int) -> List[int]:
+async def task_wait_n(n: int, max_delay: int) -> List[float]:
     """Return a sorted list of all the delays (float values) got from
     the task_wait_random function executed n times"""
     list = []
